requirements.txt
uv
ddgs
boto3
bedrock-agentcore
bedrock-agentcore-starter-toolkit
#strands-agents[otel]
elastic-opentelemetry
strands-agents-tools